Struct revlib::SparkMax

pub struct SparkMax { /* private fields */ }

Implementations

impl SparkMax

pub fn new(can_id: i32, motor_type: MotorType) -> Result<SparkMax, REVError>

pub fn new_flex( can_id: i32, motor_type: MotorType ) -> Result<SparkMax, REVError>

This should work for getting the relative encoder and setting the voltage, but it is untested. All other functionality has been ignored.

pub fn set(&mut self, speed: f32) -> Result<(), REVError>

pub fn set_smart_current_limit(&mut self, limit: u8) -> Result<(), REVError>

pub fn set_pid( &mut self, p: f32, d: f32, i: f32, feedforward: f32 ) -> Result<(), REVError>

pub fn set_pid_range( &mut self, range: RangeInclusive<f32> ) -> Result<(), REVError>

pub fn set_wrapping( &mut self, wrapping: bool, max: f32, min: f32 ) -> Result<(), REVError>

pub fn get_absolute_encoder( &mut self ) -> Result<SparkMaxAbsoluteEncoder, REVError>

pub fn set_pid_input<T: FeedbackSensor>( &mut self, sensor: &T ) -> Result<(), REVError>

pub fn reset_settings(&mut self) -> Result<(), REVError>

pub fn write_settings(&mut self) -> Result<(), REVError>

pub fn get_relative_encoder( &mut self ) -> Result<SparkMaxRelativeEncoder, REVError>

pub fn set_reference( &mut self, value: f32, control_type: ControlType ) -> Result<(), REVError>

pub fn follow(&mut self, other: &SparkMax, invert: bool) -> Result<(), REVError>

pub fn set_soft_limit( &mut self, direction: SoftLimitDirection, value: Option<f32> ) -> Result<(), REVError>

Sets and enables the limit if a value is passed, disables the limit if not value is passed